requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==4.4.2
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
plivo==4.40.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.6
Werkzeug==3.0.0
